requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
presidio-analyzer>=2.3.0
presidio-anonymizer>=2.2.0
pydantic-settings>=2.0.0
